pi/pyproject.toml
[project]
name = "gnss-gateway"
version = "0.1.0"
dependencies = [
  "fastapi",
  "uvicorn",
  "pydantic",
  "pyyaml"
]

[project.optional-dependencies]
dev = [
  "pytest",
  "ruff"
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"
